Factor actual resolver logic into a module
Begin to write a test for it
